# Unified fired event handler