--- a/addons/advanced_ballistics/CfgEventHandlers.hpp
+++ b/addons/advanced_ballistics/CfgEventHandlers.hpp
@@ -9,11 +9,3 @@ class Extended_PostInit_EventHandlers {
         init = QUOTE( call COMPILE_FILE(XEH_postInit) );
     };
 };
-
-class Extended_FiredBIS_EventHandlers {
-    class CAManBase {
-        class ADDON {
-            firedBIS = QUOTE(_this call FUNC(handleFired));
-        };
-    };
-};