--- a/addons/common/CfgEventHandlers.hpp
+++ b/addons/common/CfgEventHandlers.hpp
@@ -59,3 +59,8 @@ class Extended_Local_EventHandlers {
     };
 };
 
+class Extended_FiredBIS_EventHandlers {
+    class All {
+        ADDON = QUOTE(_this call FUNC(firedEH));
+    };
+};