--- a/addons/common/functions/fnc_getGunner.sqf
+++ b/addons/common/functions/fnc_getGunner.sqf
@@ -30,7 +30,7 @@ private _gunner = objNull;
 
 // ensure that at least the pilot is returned if there is no gunner
 if (isManualFire _vehicle && {isNull _gunner}) then {
-    _gunner = driver _vehicle;
+    _gunner = effectiveCommander _vehicle;
 };
 
 _gunner