--- a/addons/disposable/CfgEventHandlers.hpp
+++ b/addons/disposable/CfgEventHandlers.hpp
@@ -10,14 +10,6 @@ class Extended_PostInit_EventHandlers {
     };
 };
 
-class Extended_FiredBIS_EventHandlers {
-    class CAManBase {
-        class ADDON {
-            firedBIS = QUOTE(_this call FUNC(replaceATWeapon));
-        };
-    };
-};
-
 // handle preloaded missile
 class Extended_InitPost_EventHandlers {
     class CAManBase {