--- a/addons/disposable/XEH_postInit.sqf
+++ b/addons/disposable/XEH_postInit.sqf
@@ -12,3 +12,8 @@ if (!hasInterface) exitWith {};
     [_unit] call FUNC(takeLoadedATWeapon);
     [_unit] call FUNC(updateInventoryDisplay);
 }] call EFUNC(common,addEventHandler);
+
+// Register fire event handler
+// Only for the local player and for AI. Non-local players will handle it themselves
+["firedPlayer", DFUNC(replaceATWeapon)] call EFUNC(common,addEventHandler);
+["firedNonPlayer", DFUNC(replaceATWeapon)] call EFUNC(common,addEventHandler);