--- a/addons/frag/CfgEventhandlers.hpp
+++ b/addons/frag/CfgEventhandlers.hpp
@@ -9,9 +9,3 @@ class Extended_PostInit_EventHandlers {
         init = QUOTE(call COMPILE_FILE(XEH_postInit));
     };
 };
-
-class Extended_FiredBIS_EventHandlers {
-    class AllVehicles {
-        ADDON = QUOTE(_this call FUNC(fired));
-    };
-};