--- a/addons/goggles/CfgEventHandlers.hpp
+++ b/addons/goggles/CfgEventHandlers.hpp
@@ -19,14 +19,6 @@ class Extended_Killed_EventHandlers {
     };
 };
 
-class Extended_FiredBIS_EventHandlers {
-    class CAManBase {
-        class ADDON {
-            clientFiredBIS = QUOTE(if (local (_this select 0)) then {_this call FUNC(handleFired)});
-        };
-    };
-};
-
 class Extended_Explosion_EventHandlers {
     class CAManBase {
         class ADDON {