… addons/grenades/CfgEventHandlers.hpp
… addons/grenades/XEH_postInit.sqf
… addons/grenades/functions/fnc_throwGrenade.sqf
… addons/huntir/CfgEventhandlers.hpp
… addons/huntir/XEH_postInit.sqf
… addons/huntir/functions/fnc_handleFired.sqf
… addons/main/script_macros.hpp
… addons/nightvision/CfgEventHandlers.hpp
… addons/nightvision/XEH_postInitClient.sqf
… addons/nightvision/functions/fnc_blending.sqf
… addons/optics/CfgEventHandlers.hpp
… addons/optics/XEH_postInit.sqf
… addons/optics/functions/fnc_handleFired.sqf
… addons/overheating/CfgEventHandlers.hpp
… addons/overheating/XEH_postInit.sqf
… addons/overheating/functions/fnc_firedEH.sqf
… addons/overpressure/CfgEventHandlers.hpp
… addons/overpressure/XEH_postInit.sqf
… addons/overpressure/XEH_preInit.sqf
… addons/overpressure/functions/fnc_fireLauncherBackblast.sqf
… addons/overpressure/functions/fnc_fireOverpressureZone.sqf
… addons/overpressure/functions/fnc_firedEHBB.sqf
… addons/overpressure/functions/fnc_firedEHOP.sqf
… addons/recoil/CfgEventHandlers.hpp
… addons/recoil/XEH_postInit.sqf
… addons/recoil/functions/fnc_camshake.sqf
… addons/scopes/CfgEventHandlers.hpp
… addons/scopes/XEH_postInit.sqf
… addons/scopes/functions/fnc_firedEH.sqf
… addons/weaponselect/CfgEventHandlers.hpp