Optimization of fired Event Handlers

Fired Event Handlers in ACE:
- [x] advanced_ballistics
- [x] disposable
- [x] fcs
- [x] frag
- [x] goggles
- [x] grenades
- [x] huntir
- [x] nightvision
- [x] optics
- [x] overheating
- [x] overpressure
- [x] recoil
- [x] scopes
- [x] weaponselect
- [x] winddeflection

I won't do:
- [ ] javelin (I don't dare)
- [ ] missileguidance (I don't dare)
- [ ] mk6mortar (it's faster not to)