@esteldunedain 2016-02-06
Well, except for the inevitable bugs this is ready. All the listed modules should be tested. 

This should offer noticeable performance gains when AI fires. From the 16 modules, only 3 currently listen for AI fires (and one of them is default off).