@esteldunedain 2016-02-07
Ongoing testing:
- [ ] advanced_ballistics
- [x] disposable
- [x] fcs
- [ ] frag
- [ ] goggles
- [ ] grenades
- [ ] huntir
- [ ] nightvision
- [ ] optics
- [ ] overheating
- [x] overpressure
- [ ] recoil
- [ ] scopes
- [ ] weaponselect
- [x] winddeflection