@esteldunedain 2016-02-07
I completed preliminary testing of all modules and reviewed the diff. After fixing a couple of errors all typical cases of the features work (corner cases should, but they are hard to test systematically).

@PabstMirror, unfortunately the version you checked was running most event handlers, so performance should be measured again. If possible, I'd love to see the impact on AI.

@Glowbal, after performance is measured, I'd suggest this PR is merged well before 3.5.0 so any obvious bug can be found.