@esteldunedain 2016-02-09
@PabstMirror measured the following performance improvement:

## Player firing

Before:
`["B_HeavyGunner_F","Extended_FiredBISEH"]: Overall Average: 0.247898s / 390 = 0.635636ms`

After: 
`["B_HeavyGunner_F","Extended_FiredBISEH"]: Overall Average: 0.190063s / 390 = 0.487342ms`

Time saving: 23%

## AI firing

Before:
`0.286349ms`

After: 
`0.127675ms`

Time saving: 55%

Based on that, I think we may merge and test @Glowbal.